refactor rds dir to common
