+++ dcp_prototype/backend/browser/scripts/create_db.py
@@ -7,7 +7,7 @@ pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))  # 
 sys.path.insert(0, pkg_root)  # noqa
 
 from browser.code.config.db_config import BrowserDbConfig
-from browser.code.rds import browser_orm as orm
+from browser.code.common import browser_orm as orm
 
 
 if __name__ == "__main__":
